Crew Ai/AI Healthcare Blog Generator  (Powered by Gemini 1.5-flash and Crew AI)/AI Healthcare Blog Generator with Crew AI(Code)/requirements.txt
crewai
langchain_google_genai   #GOOGLE Gemini Model
load_dotenv
crewai_tools